+++ test/unit/global-hooks.js
@@ -1,8 +1,0 @@
-'use strict';
-
-var ProxyLists = require('../../index');
-
-after(function(done) {
-	if (!ProxyLists.sourcer) return done();
-	ProxyLists.sourcer.close(done);
-});