+++ test/unit/listSources.js
@@ -7,15 +7,25 @@ var ProxyLists = require('../../index');
 
 describe('listSources([options])', function() {
 
+	afterEach(function() {
+		ProxyLists._sources = [];
+	});
+
 	it('should return an array of all available sources', function() {
 
-		var sources = ProxyLists.listSources();
+		var sourceNames = ['one', 'two', 'three'];
+		_.each(sourceNames, function(name) {
+			ProxyLists.addSource(name, {
+				getProxies: _.noop,
+			});
+		});
+		var sources = ProxyLists.listSources({ sourcesDir: null });
 		expect(sources).to.be.an('array');
-		expect(sources).to.have.length(_.size(ProxyLists.sourcer.sources));
+		expect(sources).to.have.length(sourceNames.length);
 		_.each(sources, function(source) {
-			expect(_.has(source, 'name')).to.equal(true);
+			expect(source).to.have.property('name');
 			expect(source.name).to.be.a('string');
-			expect(_.has(source, 'homeUrl')).to.equal(true);
+			expect(source).to.have.property('homeUrl');
 		});
 	});
 
@@ -25,13 +35,21 @@ describe('listSources([options])', function() {
 
 			it('should return an array of only the sources in the "sourcesWhiteList"', function() {
 
+				var sourceNames = ['one', 'two', 'three'];
+				_.each(sourceNames, function(name) {
+					ProxyLists.addSource(name, {
+						getProxies: _.noop,
+					});
+				});
 				var sourcesWhiteLists = [
 					[],
-					['freeproxylists']
+					['two']
 				];
-
 				_.each(sourcesWhiteLists, function(sourcesWhiteList) {
-					var sources = ProxyLists.listSources({ sourcesWhiteList: sourcesWhiteList });
+					var sources = ProxyLists.listSources({
+						sourcesDir: null,
+						sourcesWhiteList: sourcesWhiteList,
+					});
 					expect(sources).to.be.an('array');
 					expect(sources).to.have.length(sourcesWhiteList.length);
 					_.each(sources, function(source) {
@@ -45,15 +63,23 @@ describe('listSources([options])', function() {
 
 			it('should return an array of only the sources not in the "sourcesBlackList"', function() {
 
+				var sourceNames = ['one', 'two', 'three'];
+				_.each(sourceNames, function(name) {
+					ProxyLists.addSource(name, {
+						getProxies: _.noop,
+					});
+				});
 				var sourcesBlackLists = [
 					[],
-					['freeproxylists']
+					['three']
 				];
-
 				_.each(sourcesBlackLists, function(sourcesBlackList) {
-					var sources = ProxyLists.listSources({ sourcesBlackList: sourcesBlackList });
+					var sources = ProxyLists.listSources({
+						sourcesBlackList: sourcesBlackList,
+						sourcesDir: null,
+					});
 					expect(sources).to.be.an('array');
-					expect(sources).to.have.length(_.size(ProxyLists.sourcer.sources) - sourcesBlackList.length);
+					expect(sources).to.have.length(sourceNames.length - sourcesBlackList.length);
 					_.each(sources, function(source) {
 						expect(!_.contains(sourcesBlackList, source.name)).to.equal(true);
 					});